refactor:constants 항목을 응집도를 높이기 위하여 error, statistic, validate로 분류
