+++ src/constants/statisticConstants.js
@@ -1,0 +1,15 @@
+export const MATCH_KEY = Object.freeze({
+  THREE: 3,
+  FOUR: 4,
+  FIVE: 5,
+  FIVE_AND_BONUS: 5.5,
+  SIX: 6,
+});
+
+export const MATCH_PRIZE = Object.freeze({
+  THREE: 5_000,
+  FOUR: 50_000,
+  FIVE: 1_500_000,
+  FIVE_AND_BONUS: 30_000_000,
+  SIX: 2_000_000_000,
+});
