+++ src/constants/validateConstants.js
@@ -1,0 +1,11 @@
+export const YES = "y";
+export const NO = "n";
+export const UPPER_YES = YES.toUpperCase();
+export const UPPER_NO = NO.toUpperCase();
+
+export const MIN_UNIT = 1_000;
+export const MAX_AMOUNT = 100_000;
+
+export const MIN_LOTTO_NUMBER = 1;
+export const MAX_LOTTO_NUMBER = 45;
+export const LOTTO_LENGTH = 6;
