+++ src/validations/validateWinningNumbers.js
@@ -1,9 +1,9 @@
 import {
-  WINNING_NUMBERS_ERROR_MESSAGES,
   LOTTO_LENGTH,
   MIN_LOTTO_NUMBER,
   MAX_LOTTO_NUMBER,
-} from "../constants/constants.js";
+} from "../constants/validateConstants.js";
+import { WINNING_NUMBERS_ERROR_MESSAGES } from "../constants/errorConstants.js";
 import throwIfInvalid from "../utils/throwIfInvalid.js";
 
 const parseWinningNumbers = (input) => {
@@ -17,35 +17,35 @@ const parseWinningNumbers = (input) => {
 const checkLength = (winningNumbers) => {
   throwIfInvalid(
     winningNumbers.length !== LOTTO_LENGTH,
-    WINNING_NUMBERS_ERROR_MESSAGES.INVALID_COUNT,
+    WINNING_NUMBERS_ERROR_MESSAGES.INVALID_COUNT
   );
 };
 
 const checkIsNumber = (winningNumber) => {
   throwIfInvalid(
     Number.isNaN(winningNumber),
-    WINNING_NUMBERS_ERROR_MESSAGES.NOT_A_NUMBER,
+    WINNING_NUMBERS_ERROR_MESSAGES.NOT_A_NUMBER
   );
 };
 
 const chechIsInteger = (winningNumber) => {
   throwIfInvalid(
     !Number.isInteger(winningNumber),
-    WINNING_NUMBERS_ERROR_MESSAGES.NOT_AN_INTEGER,
+    WINNING_NUMBERS_ERROR_MESSAGES.NOT_AN_INTEGER
   );
 };
 
 const checkIsInRange = (winningNumber) => {
   throwIfInvalid(
     MIN_LOTTO_NUMBER > winningNumber || MAX_LOTTO_NUMBER < winningNumber,
-    WINNING_NUMBERS_ERROR_MESSAGES.OUT_OF_RANGE,
+    WINNING_NUMBERS_ERROR_MESSAGES.OUT_OF_RANGE
   );
 };
 
 const checkIsNotDuplicate = (winningNumbers, winningNumbersSet) => {
   throwIfInvalid(
     winningNumbers.length !== winningNumbersSet.size,
-    WINNING_NUMBERS_ERROR_MESSAGES.DUPLICATE_NUMBER,
+    WINNING_NUMBERS_ERROR_MESSAGES.DUPLICATE_NUMBER
   );
 };
